:recycle: Refactor: useCalendarView FSD 적용
