+++ src/entities/calendar/model/Holiday.ts
@@ -1,0 +1,5 @@
+import { HOLIDAY_RECORD } from '../config/holidays.ts';
+
+export type HolidayRecord = typeof HOLIDAY_RECORD;
+
+export type HolidayKeys = keyof HolidayRecord;
